removing singelton class.
